compredospequenos/requirements.txt
asgiref==3.2.7
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
conditional==1.3
decorator==4.4.2
Django==3.0.6
django-appconf==1.0.4
django-autocomplete-light==3.8.1
django-autoslug==1.9.8
django-bootstrap-datepicker-plus==3.0.5
django-braces==1.14.0
django-ckeditor==6.0.0
django-clone==1.1.10
django-cors-headers==3.6.0
django-crispy-forms==1.10.0
django-crum==0.7.9
django-currentuser==0.5.2
django-extensions==3.1.0
django-filter==2.4.0
django-fontawesome-5==1.0.18
django-geojson==3.0.0
django-gravatar2==1.4.4
django-hitcount==1.3.2
django-js-asset==1.2.2
django-leaflet==0.26.0
django-redis==4.12.1
django-select2==7.5.0
django-smart-selects==1.5.9
django-taggit==1.3.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
Flask==1.1.2
future==0.18.2
GDAL==2.4.2
geocoder==1.38.1
geojson==2.5.0
geos==0.2.2
gunicorn==20.0.4
herepy==3.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
leaflet==0.0.3
lxml==4.5.1
MarkupSafe==1.1.1
Pillow==7.1.2
pkg-resources==0.0.0
ply==3.11
psycopg2-binary==2.8.5
pyasn1==0.4.8
pycryptodomex==3.9.9
pysmi==0.3.4
pysnmp==4.4.12
pytz==2020.1
ratelim==0.1.6
redis==3.5.3
requests==2.24.0
setproctitle==1.2.1
six==1.15.0
sqlparse==0.3.1
taggit-selectize==2.10.0
urllib3==1.25.11
Werkzeug==1.0.1
